Move Ole Db Tests to Their Own Test Class
